광고 보상, 활동 보상 모델 변경

+++ lib/platform/models/archive_detail_item_model.dart
@@ -18,8 +18,8 @@ class ArchiveDetailItemModel {
   String? state;
   int? badgeIssueId;
   int? challengeId;
-  int? rewardGoExerciseSum;
-  int? rewardGoAdSum;
+  double? rewardGoExerciseSum;
+  double? rewardGoAdSum;
   String? badgeName;
   String? challengeTitle;
   String? locations;
